e2e/server/requirements.txt
gunicorn==19.7.1
honcho==1.0.1
Superdesk-Core==1.30.1
-e ../../
